Rename MDC package and fields.
