+++ build-coordinator/src/main/java/org/jboss/pnc/coordinator/builder/BuildTasksInitializer.java
@@ -18,7 +18,7 @@
 
 package org.jboss.pnc.coordinator.builder;
 
-import org.jboss.pnc.common.mdc.MDCUtils;
+import org.jboss.pnc.common.logging.MDCUtils;
 import org.jboss.pnc.coordinator.builder.datastore.DatastoreAdapter;
 import org.jboss.pnc.model.BuildConfigSetRecord;
 import org.jboss.pnc.model.BuildConfiguration;
