+++ build-coordinator/src/main/java/org/jboss/pnc/coordinator/builder/DefaultBuildCoordinator.java
@@ -20,8 +20,8 @@ package org.jboss.pnc.coordinator.builder;
 import org.jboss.pnc.common.concurrent.MDCExecutors;
 import org.jboss.pnc.common.concurrent.NamedThreadFactory;
 import org.jboss.pnc.common.json.moduleconfig.SystemConfig;
-import org.jboss.pnc.common.mdc.BuildTaskContext;
-import org.jboss.pnc.common.mdc.MDCUtils;
+import org.jboss.pnc.common.logging.BuildTaskContext;
+import org.jboss.pnc.common.logging.MDCUtils;
 import org.jboss.pnc.common.monitor.PullingMonitor;
 import org.jboss.pnc.coordinator.BuildCoordinationException;
 import org.jboss.pnc.coordinator.builder.datastore.DatastoreAdapter;
