+++ common/src/main/java/org/jboss/pnc/common/logging/BuildTaskContext.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.common.mdc;
+package org.jboss.pnc.common.logging;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
