+++ rest/src/main/java/org/jboss/pnc/rest/endpoint/BuildEndpoint.java
@@ -23,8 +23,8 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
-import org.jboss.pnc.common.mdc.BuildTaskContext;
-import org.jboss.pnc.common.mdc.MDCUtils;
+import org.jboss.pnc.common.logging.BuildTaskContext;
+import org.jboss.pnc.common.logging.MDCUtils;
 import org.jboss.pnc.model.BuildRecord;
 import org.jboss.pnc.model.User;
 import org.jboss.pnc.pncmetrics.rest.TimedMetric;
