+++ rest/src/main/java/org/jboss/pnc/rest/endpoint/BuildTaskEndpoint.java
@@ -33,8 +33,8 @@ import org.jboss.pnc.common.Configuration;
 import org.jboss.pnc.common.json.moduleconfig.SystemConfig;
 import org.jboss.pnc.common.json.moduleconfig.UIModuleConfig;
 import org.jboss.pnc.common.json.moduleprovider.PncConfigProvider;
-import org.jboss.pnc.common.mdc.BuildTaskContext;
-import org.jboss.pnc.common.mdc.MDCUtils;
+import org.jboss.pnc.common.logging.BuildTaskContext;
+import org.jboss.pnc.common.logging.MDCUtils;
 import org.jboss.pnc.rest.restmodel.BuildExecutionConfigurationRest;
 import org.jboss.pnc.rest.restmodel.bpm.BuildResultRest;
 import org.jboss.pnc.rest.restmodel.response.AcceptedResponse;
