+++ spi/src/main/java/org/jboss/pnc/spi/coordinator/BuildCoordinator.java
@@ -17,7 +17,7 @@
  */
 package org.jboss.pnc.spi.coordinator;
 
-import org.jboss.pnc.common.mdc.BuildTaskContext;
+import org.jboss.pnc.common.logging.BuildTaskContext;
 import org.jboss.pnc.model.BuildConfiguration;
 import org.jboss.pnc.model.BuildConfigurationAudited;
 import org.jboss.pnc.model.BuildConfigurationSet;
